+++ packages/docs/site/src/routes/compounds/index.svelte
@@ -9,7 +9,10 @@
 		<a href='compounds/mapbox'>@svizzle/mapbox</a>
 	</li>
 	<li>
-		<a href='compounds/mapbox'>@svizzle/time_region_value</a>
+		<a href='compounds/time_region_value'>@svizzle/time_region_value</a>
+	</li>
+	<li>
+		<a href='compounds/viewports'>@svizzle/ui: viewports</a>
 	</li>
 </ul>