+++ packages/docs/site/src/routes/compounds/mapbox/all.svelte
@@ -22,7 +22,7 @@
 	}
 </script>
 
-<div class="map-container">
+<div class='map-container'>
 	<Mapbox
 		{accessToken}
 		{bounds}