+++ packages/docs/site/src/routes/compounds/mapbox/components/Layer2.svelte
@@ -34,7 +34,7 @@
 	x={cx}
 	y={cy}
 	text-anchor='middle'
-	dominant-baseline="middle"
+	dominant-baseline='middle'
 >
 	Layer 2
 </text>