+++ packages/docs/site/src/routes/compounds/mapbox/custom_controls.svelte
@@ -10,7 +10,7 @@
 	}
 </script>
 
-<div class="map-container">
+<div class='map-container'>
 	<Mapbox
 		{accessToken}
 		{bounds}