+++ CHANGELOG.md
@@ -29,11 +29,16 @@
 
 ## `@svizzle/site` v0.4.10 (next)
 
+- added examples for `@svizzle/ui`'s viewports
 - upgraded `d3-array`, `d3-geo`, `lamb`, `svelte`
 
 ## `@svizzle/ui` v0.13.0 (next)
 
 - added `Input` and `Pill` components
+- added a minimal viewport system, see:
+	- `View.svelte`
+	- `ViewsSlider.svelte`
+	- `ViewsXor.svelte`
 - updated `Banner` & `Scroller`
 - removed `just-compare`, now using `areEqual` from `@svizzle/utils`
 - upgraded: