+++ packages/docs/site/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## `@svizzle/site` v0.4.10 (next)
 
+- added examples for `@svizzle/ui`'s viewports
 - upgraded `d3-array`, `d3-geo`, `lamb`, `svelte`
 
 ## `@svizzle/site` v0.4.9